[WIP] create show results action
